requirements.txt
flask
flask-login
gunicorn
pyodbc
python-dotenv
